connected-agents/requirements.txt
azure-ai-projects
azure-identity
streamlit
python-dotenv